Adding vi-like arrow keys to right hand punctpad
The Lily58Pro keyboard doesn't has fewer keys on the 5th row than
the ErgoDox, Atreus62, and Idobo/ID75. To get used the lack of
enough keys for arrow keys on Layer 0&1, I wanted to test adding
arrow keys on Layer 2&3 for the right hand, on home row, with vi-like
controls, but shift to the right by 1 key.